Move transports to lib/transports/ directory
