+++ lib/transports/irest.js
@@ -17,7 +17,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 const http = require('http');
-const { validate } = require('./utils');
+const { validate } = require('../utils');
 
 const iRestHttp = (config, xmlInput, done) => {
   const test = validate(config);
